// src/app/components/merchandising/models/common-files.model.ts
export class CommonFiles {
    id: number;
    fileObjectId: number;
    refId: number;
    fileSize: number;
    fileType: string;
    location: string;
    uploadDate: Date;
    docTitle: string;
    version: number;
    fileVersion: number;
    activeStatus: string;
    owner: string;
    styleId:number;
    styleName: string;
    employeeName: string;
    fileName: string;
    styleFabricId: number;
    markerNo: number;
}